fixes another non-integer warning that will cause an error in the future

+++ sunpy/image/coalignment.py
@@ -452,10 +452,10 @@ def calculate_match_template_shift(mc, template=None, layer_index=0,
     nt = len(mc.maps)
 
     # Calculate a template.  If no template is passed then define one
-    # from the the index layer.
+    # from the index layer.
     if template is None:
-        tplate = mc.maps[layer_index].data[ny / 4: 3 * ny / 4,
-                                           nx / 4: 3 * nx / 4]
+        tplate = mc.maps[layer_index].data[int(ny/4): int(3*ny/4),
+                                           int(nx/4): int(3*nx/4)]
     elif isinstance(template, GenericMap):
         tplate = template.data
     elif isinstance(template, np.ndarray):
